+++ changelog.md
@@ -1,2 +1,5 @@
-# v0.2.0
-- Addition of Population Statistics
+# v0.1.4
+- Bug fixes
+- 3.51x Performance Improvement in single_point crossover
+- 2.35x Performance Improvement in two_point crossover
+- Addition of `benchmark/0` function to benchmark your algorithm